# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Medical Agency Payment Management API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface LoginInput {
  username: string;
  password: string;
}

export interface AuthResponse {
  token: string;
  username: string;
}

export interface Customer {
  id: number;
  name: string;
  ownerName: string;
  mobile: string;
  /** @nullable */
  gstNumber?: string | null;
  address: string;
  creditLimit: number;
  dueDays: number;
  totalOutstanding: number;
  createdAt: string;
}

export interface CustomerInput {
  name: string;
  ownerName: string;
  mobile: string;
  gstNumber?: string;
  address: string;
  creditLimit: number;
  dueDays: number;
}

export interface CustomerUpdate {
  name?: string;
  ownerName?: string;
  mobile?: string;
  gstNumber?: string;
  address?: string;
  creditLimit?: number;
  dueDays?: number;
}

export type InvoiceStatus = typeof InvoiceStatus[keyof typeof InvoiceStatus];


export const InvoiceStatus = {
  paid: 'paid',
  partial: 'partial',
  pending: 'pending',
  overdue: 'overdue',
} as const;

export interface Invoice {
  id: number;
  customerId: number;
  customerName: string;
  invoiceNumber: string;
  invoiceDate: string;
  billAmount: number;
  dueDate: string;
  outstandingBalance: number;
  status: InvoiceStatus;
  createdAt: string;
}

export interface InvoiceInput {
  customerId: number;
  invoiceNumber: string;
  invoiceDate: string;
  billAmount: number;
  dueDate: string;
}

export interface InvoiceUpdate {
  invoiceNumber?: string;
  invoiceDate?: string;
  billAmount?: number;
  dueDate?: string;
  status?: string;
}

export type PaymentPaymentMode = typeof PaymentPaymentMode[keyof typeof PaymentPaymentMode];


export const PaymentPaymentMode = {
  cash: 'cash',
  upi: 'upi',
  bank_transfer: 'bank_transfer',
  cheque: 'cheque',
} as const;

export interface PaymentAllocation {
  invoiceId: number;
  invoiceNumber: string;
  amount: number;
}

export interface Payment {
  id: number;
  customerId: number;
  customerName: string;
  paymentDate: string;
  amount: number;
  paymentMode: PaymentPaymentMode;
  /** @nullable */
  notes?: string | null;
  allocations?: PaymentAllocation[];
  createdAt: string;
}

export type PaymentInputPaymentMode = typeof PaymentInputPaymentMode[keyof typeof PaymentInputPaymentMode];


export const PaymentInputPaymentMode = {
  cash: 'cash',
  upi: 'upi',
  bank_transfer: 'bank_transfer',
  cheque: 'cheque',
} as const;

export interface PaymentInput {
  customerId: number;
  paymentDate: string;
  amount: number;
  paymentMode: PaymentInputPaymentMode;
  notes?: string;
}

export interface DashboardStats {
  totalOutstanding: number;
  totalPaid: number;
  todayCollection: number;
  thisMonthCollection: number;
  totalCustomers: number;
  overdueCount: number;
  dueIn3DaysCount: number;
}

export interface MonthlyCollection {
  month: number;
  year: number;
  amount: number;
  label: string;
}

export type LedgerEntryType = typeof LedgerEntryType[keyof typeof LedgerEntryType];


export const LedgerEntryType = {
  invoice: 'invoice',
  payment: 'payment',
} as const;

export interface LedgerEntry {
  date: string;
  type: LedgerEntryType;
  /** @nullable */
  invoiceNumber?: string | null;
  amount: number;
  balance: number;
  /** @nullable */
  notes?: string | null;
}

export interface OutstandingReport {
  customerId: number;
  customerName: string;
  mobile: string;
  totalOutstanding: number;
  invoiceCount: number;
  /** @nullable */
  oldestDueDate?: string | null;
}

export interface OverdueReport {
  invoiceId: number;
  invoiceNumber: string;
  customerId: number;
  customerName: string;
  mobile: string;
  billAmount: number;
  outstandingBalance: number;
  dueDate: string;
  daysOverdue: number;
}

export interface DateWiseCollection {
  date: string;
  amount: number;
  count: number;
}

export type GetCustomersParams = {
search?: string;
};

export type GetInvoicesParams = {
search?: string;
status?: string;
customerId?: number;
fromDate?: string;
toDate?: string;
};

export type GetPaymentsParams = {
customerId?: number;
fromDate?: string;
toDate?: string;
paymentMode?: string;
};

export type GetPaymentHistoryParams = {
customerId?: number;
fromDate?: string;
toDate?: string;
};

export type GetDateWiseCollectionParams = {
fromDate?: string;
toDate?: string;
};

export type GetMonthlyCollectionReportParams = {
year?: number;
};