# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Medical Agency Payment Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /customers:
    get:
      operationId: getCustomers
      tags: [customers]
      summary: List all customers
      parameters:
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of customers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Customer"
    post:
      operationId: createCustomer
      tags: [customers]
      summary: Create a new customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerInput"
      responses:
        "201":
          description: Customer created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
  /customers/{id}:
    get:
      operationId: getCustomer
      tags: [customers]
      summary: Get customer by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Customer detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
        "404":
          description: Not found
    patch:
      operationId: updateCustomer
      tags: [customers]
      summary: Update customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerUpdate"
      responses:
        "200":
          description: Updated customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
    delete:
      operationId: deleteCustomer
      tags: [customers]
      summary: Delete customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /customers/{id}/ledger:
    get:
      operationId: getCustomerLedger
      tags: [customers]
      summary: Get customer ledger
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Ledger entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LedgerEntry"
  /customers/{id}/invoices:
    get:
      operationId: getCustomerInvoices
      tags: [customers]
      summary: Get invoices for a customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Customer invoices
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Invoice"
  /invoices:
    get:
      operationId: getInvoices
      tags: [invoices]
      summary: List all invoices
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: customerId
          in: query
          schema:
            type: integer
        - name: fromDate
          in: query
          schema:
            type: string
        - name: toDate
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of invoices
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Invoice"
    post:
      operationId: createInvoice
      tags: [invoices]
      summary: Create a new invoice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvoiceInput"
      responses:
        "201":
          description: Invoice created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
  /invoices/{id}:
    get:
      operationId: getInvoice
      tags: [invoices]
      summary: Get invoice by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Invoice detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
        "404":
          description: Not found
    patch:
      operationId: updateInvoice
      tags: [invoices]
      summary: Update invoice
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvoiceUpdate"
      responses:
        "200":
          description: Updated invoice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
    delete:
      operationId: deleteInvoice
      tags: [invoices]
      summary: Delete invoice
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
...
    Customer:
      type: object
      required: [id, name, ownerName, mobile, address, creditLimit, dueDays, totalOutstanding, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        ownerName:
          type: string
        mobile:
          type: string
        gstNumber:
          type: ["string", "null"]
        address:
          type: string
        creditLimit:
          type: number
        dueDays:
          type: integer
        totalOutstanding:
          type: number
        createdAt:
          type: string
    CustomerInput:
      type: object
      required: [name, ownerName, mobile, address, creditLimit, dueDays]
      properties:
        name:
          type: string
        ownerName:
          type: string
        mobile:
          type: string
        gstNumber:
          type: string
        address:
          type: string
        creditLimit:
          type: number
        dueDays:
          type: integer
    CustomerUpdate:
      type: object
      properties:
        name:
          type: string
        ownerName:
          type: string
        mobile:
          type: string
        gstNumber:
          type: string
        address:
          type: string
        creditLimit:
          type: number
        dueDays:
          type: integer
    Invoice:
      type: object
      required: [id, customerId, customerName, invoiceNumber, invoiceDate, billAmount, dueDate, outstandingBalance, status, createdAt]
      properties:
        id:
          type: integer
        customerId:
          type: integer
        customerName:
          type: string
        invoiceNumber:
          type: string
        invoiceDate:
          type: string
        billAmount:
          type: number
        dueDate:
          type: string
        outstandingBalance:
          type: number
        status:
          type: string
          enum: [paid, partial, pending, overdue]
        createdAt:
          type: string
    InvoiceInput:
      type: object
      required: [customerId, invoiceNumber, invoiceDate, billAmount, dueDate]
      properties:
        customerId:
          type: integer
        invoiceNumber:
          type: string
        invoiceDate:
          type: string
        billAmount:
          type: number
        dueDate:
          type: string
    InvoiceUpdate:
      type: object
      properties:
        invoiceNumber:
          type: string
        invoiceDate:
          type: string
        billAmount:
          type: number
        dueDate:
          type: string
        status:
          type: string
...
    LedgerEntry:
      type: object
      required: [date, type, amount, balance]
      properties:
        date:
          type: string
        type:
          type: string
          enum: [invoice, payment]
        invoiceNumber:
          type: ["string", "null"]
        amount:
          type: number
        balance:
          type: number
        notes:
          type: ["string", "null"]
...