# lib/api-zod/src/generated/types/customerUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Medical Agency Payment Management API
 * OpenAPI spec version: 0.1.0
 */

export interface CustomerUpdate {
  name?: string;
  ownerName?: string;
  mobile?: string;
  gstNumber?: string;
  address?: string;
  creditLimit?: number;
  dueDays?: number;
}