# lib/api-zod/src/generated/types/invoiceUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Medical Agency Payment Management API
 * OpenAPI spec version: 0.1.0
 */

export interface InvoiceUpdate {
  invoiceNumber?: string;
  invoiceDate?: string;
  billAmount?: number;
  dueDate?: string;
  status?: string;
}